backend/requirements.txt
# Django Backend Requirements
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
redis==5.0.1
django-ratelimit==4.1.0
django-extensions==3.2.3
pillow==10.1.0

